move logic code to custom hook in add Pipeline index
